Added Obstacle avoidance and smooth turning
The robot now uses a P-like controller to evade the obstacle. once it
has finished and gets a high reading it moves forward for a little while
before repositioning itself.

It can either reposition itself on-point or smoothly, depending on how
severe the angle error is (see smoothTurningThreshold in the head of the
Navigator class).