Almost done, one last bug to fix and its fine
This code has one last bug to fix, which I should be able to take care
of tomorow morning.:
in the second case:
The angle Theta goes below -180 in some scenarios.

The robot successfully avoids obstacles but fails to get to final
destination because a problem with the direction (probably the angle)
makes it miss the radius.